requirements.txt
numpy
Pillow
scikit-learn
joblib
fastapi
uvicorn
python-multipart
streamlit
matplotlib
pandas
requests
locust
